+++ AndroidTouchProject/Assets/IControllable.cs
@@ -1,6 +1,9 @@
-﻿using UnityEngine;
+﻿using System.Collections.Generic;
+using UnityEngine;
 
 public interface IControllable
 {
     bool selectToggle();
+
+    void drag(List<Vector2> positions);
 }
