+++ AndroidTouchProject/Assets/ITouchController.cs
@@ -1,10 +1,11 @@
-﻿using UnityEngine;
+﻿using System.Collections.Generic;
+using UnityEngine;
 
 public interface ITouchController
 {
     void tap(Vector2 position);
 
-    void drag(Vector2 current_position);
+    void drag(List<Vector2> current_position);
 
     void pinch(Vector2 position1, Vector2 position2, float relative_distance);
 }
