+++ AndroidTouchProject/Assets/New Folder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e89fa796ac947b3429b14edb9c8727e4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
