+++ AndroidTouchProject/Assets/New Folder/CubeControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5da1d77efa88c3946ab072082130f4b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
