+++ AndroidTouchProject/Assets/New Folder/CubeControl2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 928d0ce93bec97f44858be23c9f65b61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
